Move util to lib/util
