+++ core/lib/util/capability.h
