+++ core/lib/util/daemon.c
