+++ core/lib/util/daemon.h
