+++ core/lib/util/message.c
