+++ core/lib/util/privilege.c
