+++ core/lib/util/privilege.h
