+++ core/lib/util/registry.c
