+++ core/lib/util/registry.h
