+++ core/lib/util/sessiondir.c
