+++ core/lib/util/sessiondir.h
