+++ core/lib/util/signal.c
