+++ core/lib/util/signal.h
