+++ core/lib/util/util.c
